// src/interfaces/address.interface.ts
interface IAddress {
  zip_code: string;
  country: string;
  state: string;
  city: string;
  district: string;
  street: string;
  number: number;
  complement: string;
}

interface IAddressUpdate {
  zip_code?: string;
  country?: string;
  state?: string;
  city?: string;
  district?: string;
  street?: string;
  number?: number;
  complement?: string;
}

export { IAddress, IAddressUpdate };